chat-websocket/requirements.txt
websockets
fastapi
uvicorn
python-jose[cryptography]
redis
